# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JIVAN DAAN Blood Donation Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  donor: "donor",
  receiver: "receiver",
  hospital: "hospital",
} as const;

export interface User {
  id: number;
  uid: string;
  name: string;
  email: string;
  role: UserRole;
  bloodGroup: string;
  location: string;
  pinCode?: string;
  image?: string | null;
  hospitalName?: string | null;
  contact?: string | null;
  isAvailable: boolean;
  createdAt: string;
}

export type CreateUserInputRole =
  (typeof CreateUserInputRole)[keyof typeof CreateUserInputRole];

export const CreateUserInputRole = {
  donor: "donor",
  receiver: "receiver",
  hospital: "hospital",
} as const;

export interface CreateUserInput {
  uid: string;
  name: string;
  email: string;
  role: CreateUserInputRole;
  bloodGroup: string;
  location: string;
  pinCode?: string;
  image?: string | null;
  hospitalName?: string | null;
  contact?: string | null;
}

export interface UpdateUserInput {
  name?: string;
  bloodGroup?: string;
  location?: string;
  pinCode?: string;
  image?: string | null;
  hospitalName?: string | null;
  contact?: string | null;
  isAvailable?: boolean;
}

export type BloodRequestUrgency =
  (typeof BloodRequestUrgency)[keyof typeof BloodRequestUrgency];

export const BloodRequestUrgency = {
  critical: "critical",
  urgent: "urgent",
  normal: "normal",
} as const;

export type BloodRequestStatus =
  (typeof BloodRequestStatus)[keyof typeof BloodRequestStatus];

export const BloodRequestStatus = {
  active: "active",
  fulfilled: "fulfilled",
  cancelled: "cancelled",
} as const;

export interface BloodRequest {
  id: number;
  bloodGroup: string;
  location: string;
  urgency: BloodRequestUrgency;
  patientName: string;
  units: number;
  contactNumber: string;
  description?: string | null;
  createdBy: string;
  createdByName: string;
  createdAt: string;
  status: BloodRequestStatus;
}

export type CreateBloodRequestInputUrgency =
  (typeof CreateBloodRequestInputUrgency)[keyof typeof CreateBloodRequestInputUrgency];

export const CreateBloodRequestInputUrgency = {
  critical: "critical",
  urgent: "urgent",
  normal: "normal",
} as const;

export interface CreateBloodRequestInput {
  bloodGroup: string;
  location: string;
  urgency: CreateBloodRequestInputUrgency;
  patientName: string;
  units: number;
  contactNumber: string;
  description?: string | null;
  createdBy: string;
  createdByName: string;
}

export interface Hospital {
  id: number;
  uid: string;
  name: string;
  location: string;
  contact: string;
  email: string;
  createdAt: string;
}

export interface CreateHospitalInput {
  uid: string;
  name: string;
  location: string;
  contact: string;
  email: string;
}

export interface PlatformStats {
  totalDonors: number;
  totalRequests: number;
  totalHospitals: number;
  livesSaved: number;
}

export type GetDonorsParams = {
  bloodGroup?: string;
  location?: string;
  pinCode?: string;
};

export type GetBloodRequestsParams = {
  bloodGroup?: string;
  urgency?: string;
};