# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JIVAN DAAN Blood Donation Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User operations
  - name: donors
    description: Donor operations
  - name: requests
    description: Blood request operations
  - name: hospitals
    description: Hospital operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create or update a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "200":
          description: User created/updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{uid}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile by UID
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /donors:
    get:
      operationId: getDonors
      tags: [donors]
      summary: Get list of donors with optional filters
      parameters:
        - name: bloodGroup
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: pinCode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of donors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /requests:
    get:
      operationId: getBloodRequests
      tags: [requests]
      summary: Get all blood requests
      parameters:
        - name: bloodGroup
          in: query
          schema:
            type: string
        - name: urgency
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of blood requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodRequest"
    post:
      operationId: createBloodRequest
      tags: [requests]
      summary: Create a new blood request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBloodRequestInput"
      responses:
        "201":
          description: Blood request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"

  /requests/{id}:
    delete:
      operationId: deleteBloodRequest
      tags: [requests]
      summary: Delete a blood request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /hospitals:
    get:
      operationId: getHospitals
      tags: [hospitals]
      summary: Get list of hospitals
      responses:
        "200":
          description: List of hospitals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hospital"
    post:
      operationId: createHospital
      tags: [hospitals]
      summary: Register a hospital
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHospitalInput"
      responses:
        "201":
          description: Hospital created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"

  /stats:
    get:
      operationId: getStats
      tags: [health]
      summary: Get platform statistics
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    User:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [donor, receiver, hospital]
        bloodGroup:
          type: string
        location:
          type: string
        pinCode:
          type: string
        image:
          type: string
          nullable: true
        hospitalName:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        isAvailable:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - uid
        - name
        - email
        - role
        - bloodGroup
        - location
        - isAvailable
        - createdAt

    CreateUserInput:
      type: object
      properties:
        uid:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [donor, receiver, hospital]
        bloodGroup:
          type: string
        location:
          type: string
        pinCode:
          type: string
        image:
          type: string
          nullable: true
        hospitalName:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
      required:
        - uid
        - name
        - email
        - role
        - bloodGroup
        - location

    UpdateUserInput:
      type: object
      properties:
        name:
          type: string
        bloodGroup:
          type: string
        location:
          type: string
        pinCode:
          type: string
        image:
          type: string
          nullable: true
        hospitalName:
          type: string
          nullable: true
        contact:
          type: string
          nullable: true
        isAvailable:
          type: boolean

    BloodRequest:
      type: object
      properties:
        id:
          type: integer
        bloodGroup:
          type: string
        location:
          type: string
        urgency:
          type: string
          enum: [critical, urgent, normal]
        patientName:
          type: string
        units:
          type: integer
        contactNumber:
          type: string
        description:
          type: string
          nullable: true
        createdBy:
          type: string
        createdByName:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, fulfilled, cancelled]
      required:
        - id
        - bloodGroup
        - location
        - urgency
        - patientName
        - units
        - contactNumber
        - createdBy
        - createdByName
        - createdAt
        - status

    CreateBloodRequestInput:
      type: object
      properties:
        bloodGroup:
          type: string
        location:
          type: string
        urgency:
          type: string
          enum: [critical, urgent, normal]
        patientName:
          type: string
        units:
          type: integer
        contactNumber:
          type: string
        description:
          type: string
          nullable: true
        createdBy:
          type: string
        createdByName:
          type: string
      required:
        - bloodGroup
        - location
        - urgency
        - patientName
        - units
        - contactNumber
        - createdBy
        - createdByName

    Hospital:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: string
        name:
          type: string
        location:
          type: string
        contact:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - uid
        - name
        - location
        - contact
        - email
        - createdAt

    CreateHospitalInput:
      type: object
      properties:
        uid:
          type: string
        name:
          type: string
        location:
          type: string
        contact:
          type: string
        email:
          type: string
      required:
        - uid
        - name
        - location
        - contact
        - email

    PlatformStats:
      type: object
      properties:
        totalDonors:
          type: integer
        totalRequests:
          type: integer
        totalHospitals:
          type: integer
        livesSaved:
          type: integer
      required:
        - totalDonors
        - totalRequests
        - totalHospitals
        - livesSaved