# lib/api-zod/src/generated/types/bloodRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JIVAN DAAN Blood Donation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BloodRequestStatus } from "./bloodRequestStatus";
import type { BloodRequestUrgency } from "./bloodRequestUrgency";

export interface BloodRequest {
  id: number;
  bloodGroup: string;
  location: string;
  urgency: BloodRequestUrgency;
  patientName: string;
  units: number;
  contactNumber: string;
  description?: string | null;
  createdBy: string;
  createdByName: string;
  createdAt: Date;
  status: BloodRequestStatus;
}