# lib/api-zod/src/generated/types/createUserInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JIVAN DAAN Blood Donation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserInputRole } from "./createUserInputRole";

export interface CreateUserInput {
  uid: string;
  name: string;
  email: string;
  role: CreateUserInputRole;
  bloodGroup: string;
  location: string;
  pinCode?: string;
  image?: string | null;
  hospitalName?: string | null;
  contact?: string | null;
}