# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JIVAN DAAN Blood Donation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalDonors: number;
  totalRequests: number;
  totalHospitals: number;
  livesSaved: number;
}