# lib/api-zod/src/generated/types/updateUserInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JIVAN DAAN Blood Donation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateUserInput {
  name?: string;
  bloodGroup?: string;
  location?: string;
  pinCode?: string;
  image?: string | null;
  hospitalName?: string | null;
  contact?: string | null;
  isAvailable?: boolean;
}