# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JIVAN DAAN Blood Donation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  uid: string;
  name: string;
  email: string;
  role: UserRole;
  bloodGroup: string;
  location: string;
  pinCode?: string;
  image?: string | null;
  hospitalName?: string | null;
  contact?: string | null;
  isAvailable: boolean;
  createdAt: Date;
}